Added option install analytics(lyx) portion seperately

+++ install.sh
@@ -1,30 +1,50 @@
 
 install_os_deps() {
-	for pkg in libuuid uuid httpd mod_wsgi python-setuptools; do
+	echo "-- Installing OS dependencies...."
+	for pkg in libuuid uuid httpd mod_wsgi python-setuptools mongodb; do
 		rpm -qa | grep ${pkg} || yum -y install ${pkg};
 	done;
 	chkconfig httpd on;
 };
 install_pydeps() {
+	echo "-- Installing python dependencies..."
 	which pip || easy_install pip;
-	for pypkg in uuid Django djangorestframework django-filter; do
+	for pypkg in uuid Django djangorestframework django-filter pymongo celery; do
 		pip list | grep ${pypkg} || pip install ${pypkg};
 	done;
 };
+install_deps() {
+	install_os_deps;
+	install_pydeps;
+}
 clean() {
 	find . -name '*.pyc' -exec rm -rvf '{}' \;
 }
 install_app(){
 	clean;
+	echo "-- Installing app..."
 	/etc/init.d/httpd stop;
 	if [ -d /opt/metrilyx ]; then
 		mv /opt/metrilyx /opt/metrilyx-$(date '+%d%b%Y_%H%M');
 	fi;
 	mkdir -p /opt/metrilyx;
 	cp -a . /opt/metrilyx/;
+	cp -a etc/rc.d/init.d/* /etc/rc.d/init.d/;
+	if [ ! -f /etc/sysconfig/celeryd ]; then 
+		cp etc/sysconfig/celeryd /etc/sysconfig/;
+	fi
+	if [ ! -f /opt/metrilyx/etc/metrilyx/metrilyx.conf ]; then
+		cp etc/metrilyx/metrilyx.conf.sample /opt/metrilyx/etc/metrilyx/metrilyx.conf;
+	fi
+}
+install_web_ui() {
+	echo "-- Install UI..."
 	cp etc/httpd/conf.d/metrilyx.conf /etc/httpd/conf.d/;
-	cp etc/metrilyx/metrilyx.conf.sample /opt/metrilyx/etc/metrilyx/metrilyx.conf;
 	chown -R apache:apache /opt/metrilyx;
+	echo "-- Setup configuration"
+	vi /opt/metrilyx/etc/metrilyx/metrilyx.conf
+	echo "-- Restarting apache..."
+	/etc/init.d/httpd restart
 }
 
 ##### Main ####
@@ -34,18 +54,17 @@ if [ ! -f "/etc/redhat-release" ]; then
 	exit 1;
 fi
 
-echo "-- Installing OS dependencies...."
-install_os_deps;
-
-echo "-- Installing python dependencies..."
-install_pydeps;
-
-echo "-- Installing..."
-install_app;
-
-echo "-- Setup configuration"
-vi /opt/metrilyx/etc/metrilyx/metrilyx.conf
-
-echo "-- Restarting apache..."
-/etc/init.d/httpd restart
-
+case "$1" in
+	lyx)	
+		install_deps;
+		install_app;
+		;;
+	all)
+		install_deps;
+		install_app;
+		install_web_ui;
+		;;	
+	*)
+		echo -e "\n\tUsage:\n\t\t$0\t[lyx|www|all]\n";
+		exit 2;
+esac
